Write a halfway decent example.

+++ src/qkl/client.c
@@ -2,10 +2,8 @@
 #include "qkl/usr/platform.h"
 #include "qkl/encode.h"
 
-// void qkl_lr_buff_init(qkl_lr_buff *buff, void *data, int data_size, int elem_size, int num_elems);
-
 void qkl_client_init(qkl_client *client, void *data, int data_size, int num_elems){
-    QKL_ASSERT(num_elems * sizeof(qkl_entry) == data_size);
+    QKL_ASSERT(num_elems * (int)sizeof(qkl_entry) == data_size);
     qkl_lr_buff_init(&client->buff, data, data_size, sizeof(qkl_entry), num_elems);
     client->dropped = 0;
 }
@@ -29,7 +27,6 @@ void qkl_client_new_entry(qkl_client *client, const char *fmt, uint8_t *data, in
         return;
     }
     ent->fmt = fmt;
-    // memcpy(ent->data, data, len);
 
     for(int i=0; i<len; i++){
         uint8_t v = data[i];
